fix(Estimate): Refactored Invoice Back into the business namespace
