+++ src/Response/Estimate/EstimateResponse.php
@@ -16,54 +16,54 @@ use ZEROSPAM\Freshbooks\Business\Enums\Currency\CurrencyEnum;
 use ZEROSPAM\Freshbooks\Business\Enums\Estimate\StatusEnum;
 use ZEROSPAM\Freshbooks\Business\Enums\Estimate\UIStatusEnum;
 use ZEROSPAM\Freshbooks\Business\Enums\Language\LanguageEnum;
-use ZEROSPAM\Freshbooks\Response\Invoice\InvoiceLine;
+use ZEROSPAM\Freshbooks\Business\InvoiceLine;
 
 /**
  * Class EstimateResponse
  *
- * @property-read string                    $province
- * @property-read string                    $code
- * @property-read Carbon                    $create_date
- * @property-read UIStatusEnum              $display_status
- * @property-read bool                      $require_client_signature
- * @property-read string                    $street
- * @property-read string|null               $vat_number
- * @property-read int                       $ownerid
- * @property-read int                       $id
- * @property-read bool                      $invoiced
- * @property-read string                    $city
- * @property-read string                    $lname
- * @property-read int                       $ext_archive
- * @property-read string                    $fname
- * @property-read int                       $vis_state
- * @property-read string                    $current_organization
- * @property-read StatusEnum                $status
- * @property-read string                    $estimate_number
- * @property-read Carbon                    $updated
- * @property-read string                    $terms
- * @property-read string                    $description
- * @property-read string|null               $vat_name
- * @property-read string                    $street2
- * @property-read string                    $template
- * @property-read UIStatusEnum              $ui_status
- * @property-read Amount                    $discount_total
- * @property-read string                    $address
- * @property-read bool                      $accepted
- * @property-read int                       $customerid
- * @property-read string                    $accounting_systemid
- * @property-read string                    $organization
- * @property-read LanguageEnum              $language
- * @property-read string                    $po_number
- * @property-read string                    $country
- * @property-read string                    $notes
- * @property-read string|null               $reply_status
- * @property-read Amount                    $amount
- * @property-read int                       $estimateid
- * @property-read int                       $sentid
- * @property-read string                    $discount_value
- * @property-read bool                      $rich_proposal
- * @property-read Carbon                    $created_at
- * @property-read CurrencyEnum              $currency_code
+ * @property-read string                   $province
+ * @property-read string                   $code
+ * @property-read Carbon                   $create_date
+ * @property-read UIStatusEnum             $display_status
+ * @property-read bool                     $require_client_signature
+ * @property-read string                   $street
+ * @property-read string|null              $vat_number
+ * @property-read int                      $ownerid
+ * @property-read int                      $id
+ * @property-read bool                     $invoiced
+ * @property-read string                   $city
+ * @property-read string                   $lname
+ * @property-read int                      $ext_archive
+ * @property-read string                   $fname
+ * @property-read int                      $vis_state
+ * @property-read string                   $current_organization
+ * @property-read StatusEnum               $status
+ * @property-read string                   $estimate_number
+ * @property-read Carbon                   $updated
+ * @property-read string                   $terms
+ * @property-read string                   $description
+ * @property-read string|null              $vat_name
+ * @property-read string                   $street2
+ * @property-read string                   $template
+ * @property-read UIStatusEnum             $ui_status
+ * @property-read Amount                   $discount_total
+ * @property-read string                   $address
+ * @property-read bool                     $accepted
+ * @property-read int                      $customerid
+ * @property-read string                   $accounting_systemid
+ * @property-read string                   $organization
+ * @property-read LanguageEnum             $language
+ * @property-read string                   $po_number
+ * @property-read string                   $country
+ * @property-read string                   $notes
+ * @property-read string|null              $reply_status
+ * @property-read Amount                   $amount
+ * @property-read int                      $estimateid
+ * @property-read int                      $sentid
+ * @property-read string                   $discount_value
+ * @property-read bool                     $rich_proposal
+ * @property-read Carbon                   $created_at
+ * @property-read CurrencyEnum             $currency_code
  * @property-read InvoiceLine[]|array|null $lines
  *
  * @package ZEROSPAM\Freshbooks\Response\Estimate
